Moved business logic to PanZoomApi service
