+++ projects/demo/src/app/app.component.scss
@@ -1,0 +1,19 @@
+.map {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  background-color: black;
+}
+
+.row {
+  display: flex;
+}
+
+
+.column {
+  flex: 1;
+  width: 150px;
+  height: 150px;
+  background-color: beige;
+  border: 1px solid black;
+}
