+++ projects/demo/src/styles.scss
@@ -1,0 +1,1 @@
+/* You can add global styles to this file, and also import other style files */
