--- projects/ngx-panzoom/src/lib/panzoom-model.ts
@@ -1,7 +1,0 @@
-import { Point } from './types/point';
-
-export interface PanZoomModel {
-  zoomLevel: number;
-  isPanning?: boolean;
-  pan: Point;
-}
