+++ projects/ngx-panzoom/src/lib/panzoom.component.html
@@ -19,6 +19,6 @@
 </div>
 
 <div
-  #panzoomOverlay
+  #overlayElement
   class="pan-zoom-overlay">
 </div>
