+++ projects/ngx-panzoom/src/lib/panzoom.module.ts
@@ -1,10 +1,11 @@
+import { PanZoomApi } from './panzoom.api';
 import { NgModule } from '@angular/core';
 import { PanZoomComponent } from './panzoom.component';
 
 @NgModule({
   imports: [],
   declarations: [ PanZoomComponent ],
-  providers: [],
+  providers: [ PanZoomApi ],
   exports: [ PanZoomComponent ]
 })
