+++ projects/ngx-panzoom/src/lib/panzoomwidget.css
@@ -6,35 +6,45 @@
   -khtml-user-select: none;
   -moz-user-select: none;
   -ms-user-select: none;
-  user-select: none; }
-  .panzoomwidget .zoom-button {
-    text-align: center;
-    height: 20px;
-    width: 100%;
-    background-color: white;
-    cursor: hand;
-    border: 1px solid black;
-    -webkit-border-radius: 2px;
-    -moz-border-radius: 2px;
-    border-radius: 2px;
-    -webkit-box-shadow: 2px 2px 2px 0 rgba(0, 0, 0, 0.2);
-    box-shadow: 2px 2px 2px 0 rgba(0, 0, 0, 0.2); }
-  .panzoomwidget .zoom-button:hover {
-    background-color: #ccc; }
-  .panzoomwidget .zoom-level {
-    width: 100%;
-    border-left: 2px solid black;
-    margin-left: 10px; }
-  .panzoomwidget .zoom-slider {
-    position: relative;
-    cursor: hand; }
-  .panzoomwidget .zoom-slider-widget {
-    position: absolute;
-    width: 20px;
-    background-color: white;
-    border: 1px solid black;
-    -webkit-border-radius: 2px;
-    -moz-border-radius: 2px;
-    -webkit-box-shadow: 2px 2px 2px 0 rgba(0, 0, 0, 0.5);
-    box-shadow: 2px 2px 2px 0 rgba(0, 0, 0, 0.2); }
+  user-select: none;
+}
+
+.panzoomwidget .zoom-button {
+  text-align: center;
+  height: 20px;
+  width: 100%;
+  background-color: white;
+  cursor: hand;
+  border: 1px solid black;
+  -webkit-border-radius: 2px;
+  -moz-border-radius: 2px;
+  border-radius: 2px;
+  -webkit-box-shadow: 2px 2px 2px 0 rgba(0, 0, 0, 0.2);
+  box-shadow: 2px 2px 2px 0 rgba(0, 0, 0, 0.2);
+}
+
+.panzoomwidget .zoom-button:hover {
+  background-color: #ccc;
+}
+
+.panzoomwidget .zoom-level {
+  width: 100%;
+  border-left: 2px solid black;
+  margin-left: 10px;
+}
+
+.panzoomwidget .zoom-slider {
+  position: relative;
+  cursor: hand;
+}
+
+.panzoomwidget .zoom-slider-widget {
+  position: absolute;
+  width: 20px;
+  background-color: white;
+  border: 1px solid black;
+  border-radius: 2px;
+  -webkit-box-shadow: 2px 2px 2px 0 rgba(0, 0, 0, 0.5);
+  box-shadow: 2px 2px 2px 0 rgba(0, 0, 0, 0.2);
+}
