--- projects/ngx-panzoom/src/lib/types/offset.ts
@@ -1,4 +1,0 @@
-export interface Offset {
-  top: number;
-  left: number;
-}
