--- projects/ngx-panzoom/src/lib/types/point.ts
@@ -1,4 +1,0 @@
-export interface Point {
-  x: number;
-  y: number;
-}
