--- projects/ngx-panzoom/src/lib/types/rect.ts
@@ -1,6 +1,0 @@
-export interface Rect {
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-}
