+++ projects/ngx-panzoom/src/public-api.ts
@@ -1,10 +1,5 @@
-/*
- * Public API Surface of ngx-panzoom
- */
-
 export * from './lib/panzoom.module';
 export * from './lib/panzoom.component';
-export * from './lib/panzoom-api';
-export * from './lib/panzoom-model';
+export * from './lib/panzoom.api';
 export * from './lib/panzoom-config';
-export * from './lib/types/panzoom-config-options';
+export * from './lib/interfaces';
