+++ tsconfig.json
@@ -9,6 +9,12 @@
   "references": [
     {
       "path": "./projects/ngx-panzoom/tsconfig.lib.json"
+    },
+    {
+      "path": "./projects/demo/tsconfig.app.json"
+    },
+    {
+      "path": "./projects/demo/tsconfig.spec.json"
     }
   ]
 }
